[client]  * Adding more definition to the progress bar

+++ client/src/es/deusto/weblab/client/ui/widgets/WlPredictiveProgressBar.java
@@ -340,7 +340,7 @@ public class WlPredictiveProgressBar extends GlProgressBar implements IWlWidget 
 		this.running = true;
 		this.setMaxProgress(this.resolution);
 		this.setMinProgress(0);
-		this.updateTime = this.estimatedTime / this.resolution;
+		this.updateTime = (int)((double)this.estimatedTime / this.resolution);
 
 		
 		this.currentTimer = new Timer(){
